Split login manager into internal modules
